+++ analysis/import/GitLogParser/src/main/kotlin/de/maibornwolff/codecharta/importer/gitlogparser/GitLogParser.kt
@@ -6,14 +6,13 @@ import de.maibornwolff.codecharta.importer.gitlogparser.converter.ProjectConvert
 import de.maibornwolff.codecharta.importer.gitlogparser.input.metrics.MetricsFactory
 import de.maibornwolff.codecharta.importer.gitlogparser.parser.LogParserStrategy
 import de.maibornwolff.codecharta.importer.gitlogparser.parser.git.GitLogNumstatRawParserStrategy
+import de.maibornwolff.codecharta.importer.gitlogparser.subcommands.LogScanCommand
+import de.maibornwolff.codecharta.importer.gitlogparser.subcommands.RepoScanCommand
 import de.maibornwolff.codecharta.model.Project
 import de.maibornwolff.codecharta.serialization.ProjectDeserializer
 import de.maibornwolff.codecharta.serialization.ProjectSerializer
 import de.maibornwolff.codecharta.tools.interactiveparser.InteractiveParser
 import de.maibornwolff.codecharta.tools.interactiveparser.ParserDialogInterface
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.runBlocking
 import org.mozilla.universalchardet.UniversalDetector
 import picocli.CommandLine
 import java.io.File
@@ -27,8 +26,9 @@ import java.util.stream.Stream
 
 @CommandLine.Command(
     name = "gitlogparser",
-    description = ["git log parser - generates cc.json from git log file"],
-    footer = ["Copyright(c) 2020, MaibornWolff GmbH"]
+    description = ["git log parser - generates cc.json from git-log files"],
+    subcommands = [LogScanCommand::class, RepoScanCommand::class],
+    footer = ["Copyright(c) 2022, MaibornWolff GmbH"]
 )
 class GitLogParser(
     private val input: InputStream = System.`in`,
@@ -36,42 +36,10 @@ class GitLogParser(
     private val error: PrintStream = System.err
 ) : Callable<Void>, InteractiveParser {
 
+    private val inputFormatNames = GIT_LOG_NUMSTAT_RAW_REVERSED
+
     @CommandLine.Option(names = ["-h", "--help"], usageHelp = true, description = ["displays this help and exits"])
     private var help = false
-
-    @CommandLine.Parameters(arity = "1", paramLabel = "FILE", description = ["file to parse"])
-    private var file: File? = null
-
-    @CommandLine.Option(
-        names = ["-n", "--file-name-list"],
-        arity = "1",
-        paramLabel = "FILE",
-        description = ["list of all file names in current git project"]
-    )
-    private var nameFile: File? = null
-
-    @CommandLine.Option(names = ["-o", "--output-file"], description = ["output File"])
-    private var outputFile: String? = null
-
-    @CommandLine.Option(
-        names = ["-nc", "--not-compressed"],
-        description = ["save uncompressed output File"],
-        arity = "0"
-    )
-    private var compress = true
-
-    @CommandLine.Option(names = ["--silent"], description = ["suppress command line output during process"])
-    private var silent = false
-
-    @CommandLine.Option(
-        names = ["--input-format"],
-        description = ["input format for parsing, optional only one type is currently supported"],
-        defaultValue = "GIT_LOG_NUMSTAT_RAW_REVERSED"
-    )
-    private var inputFormatNames: InputFormatNames = GIT_LOG_NUMSTAT_RAW_REVERSED
-
-    @CommandLine.Option(names = ["--add-author"], description = ["add an array of authors to every file"])
-    private var addAuthor = false
 
     private val logParserStrategy: LogParserStrategy
         get() = getLogParserStrategyByInputFormat(inputFormatNames)
@@ -97,11 +65,21 @@ class GitLogParser(
 
     @Throws(IOException::class)
     override fun call(): Void? {
-
         print(" ")
+        return null
+    }
+
+    internal fun buildProject(
+        gitLogFile: File,
+        gitLsFile: File,
+        outputFilePath: String?,
+        addAuthor: Boolean,
+        silent: Boolean,
+        compress: Boolean
+    ) {
         var project = createProjectFromLog(
-            file!!,
-            nameFile!!,
+            gitLogFile,
+            gitLsFile,
             logParserStrategy,
             metricsFactory,
             addAuthor,
@@ -113,9 +91,7 @@ class GitLogParser(
             project = MergeFilter.mergePipedWithCurrentProject(pipedProject, project)
         }
 
-        ProjectSerializer.serializeToFileOrStream(project, outputFile, output, compress)
-
-        return null
+        ProjectSerializer.serializeToFileOrStream(project, outputFilePath, output, compress)
     }
 
     private fun getLogParserStrategyByInputFormat(formatName: InputFormatNames): LogParserStrategy {
@@ -134,56 +110,23 @@ class GitLogParser(
     }
 
     private fun createProjectFromLog(
-        pathToLog: File,
-        pathToNameTree: File,
+        gitLogFile: File,
+        gitLsFile: File,
         parserStrategy: LogParserStrategy,
         metricsFactory: MetricsFactory,
         containsAuthors: Boolean,
         silent: Boolean = false
     ): Project {
-        val namesInProject = readFileNameListFile(pathToNameTree)
-        val encoding = guessEncoding(pathToLog) ?: "UTF-8"
+        val namesInProject = readFileNameListFile(gitLsFile)
+        val encoding = guessEncoding(gitLogFile) ?: "UTF-8"
         if (!silent) error.println("Assumed encoding $encoding")
-        val lines: Stream<String> = Files.lines(pathToLog.toPath(), Charset.forName(encoding))
+        val lines: Stream<String> = Files.lines(gitLogFile.toPath(), Charset.forName(encoding))
         val projectConverter = ProjectConverter(containsAuthors)
-        val logSizeInByte = file!!.length()
+        val logSizeInByte = gitLogFile.length()
         return GitLogProjectCreator(parserStrategy, metricsFactory, projectConverter, logSizeInByte, silent).parse(
             lines,
             namesInProject
         )
-    }
-
-    // not implemented yet #738
-    private fun printUsage() {
-        println("----")
-        printLogCreation()
-
-        println("----")
-        printMetricInfo()
-    }
-
-    private fun printLogCreation() {
-        println("  Log creation via:")
-
-        printLogCreationByInputFormatNames(inputFormatNames)
-    }
-
-    private fun printLogCreationByInputFormatNames(actualInfoFormatName: InputFormatNames?) {
-        val creationCommand = getLogParserStrategyByInputFormat(actualInfoFormatName!!).creationCommand()
-        println(String.format("  \t%s :\t\"%s\".", actualInfoFormatName, creationCommand))
-    }
-
-    private fun printMetricInfo() {
-        val infoFormat = "  \t%s:\t %s"
-        println("  Available metrics:")
-        runBlocking(Dispatchers.Default) {
-            metricsFactory.createMetrics()
-                .forEach {
-                    launch {
-                        println(String.format(infoFormat, it.metricName(), it.description()))
-                    }
-                }
-        }
     }
 
     companion object {