+++ analysis/import/GitLogParser/src/test/kotlin/de/maibornwolff/codecharta/importer/gitlogparser/parser/git/helper/GitLogNumstatParsingHelperRenamingTest.kt
@@ -2,47 +2,37 @@ package de.maibornwolff.codecharta.importer.gitlogparser.parser.git.helper
 
 import de.maibornwolff.codecharta.importer.gitlogparser.input.Modification
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.junit.runners.Parameterized
-import java.util.Arrays
+import org.junit.jupiter.api.Assertions.assertTrue
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.Arguments
+import org.junit.jupiter.params.provider.MethodSource
 import java.util.function.Function
 
-@RunWith(Parameterized::class)
-class GitLogNumstatParsingHelperRenamingTest(
-    private val fileLine: String,
-    private val oldFilename: String,
-    private val newFilename: String
-) {
+class GitLogNumstatParsingHelperRenamingTest {
 
     companion object {
-
         @JvmStatic
-        @Parameterized.Parameters(name = "{index}: {0}")
-        fun data(): Collection<Array<Any>> {
-            return Arrays.asList(
-                arrayOf<Any>(
-                    "1 2 src/{RenameOld.java => RenameNew.java}",
-                    "src/RenameOld.java",
-                    "src/RenameNew.java"
-                ),
-                arrayOf<Any>("1 2 {old => new}/Rename.java", "old/Rename.java", "new/Rename.java"),
-                arrayOf<Any>("1 2 src/{ => new}/Rename.java", "src/Rename.java", "src/new/Rename.java"),
-                arrayOf<Any>("1 2 src/{old => }/Rename.java", "src/old/Rename.java", "src/Rename.java"),
-                arrayOf<Any>("1 2 src/{old => new}/Rename.java", "src/old/Rename.java", "src/new/Rename.java"),
-                arrayOf<Any>("1\t2\tRename.java => new/Rename.java", "Rename.java", "new/Rename.java"),
-                arrayOf<Any>("1\t2\tRenameOld.java => RenameNew.java", "RenameOld.java", "RenameNew.java")
+        fun data() =
+            listOf(
+                Arguments.of("1 2 src/{RenameOld.java => RenameNew.java}", "src/RenameOld.java", "src/RenameNew.java"),
+                Arguments.of("1 2 {old => new}/Rename.java", "old/Rename.java", "new/Rename.java"),
+                Arguments.of("1 2 src/{ => new}/Rename.java", "src/Rename.java", "src/new/Rename.java"),
+                Arguments.of("1 2 src/{old => }/Rename.java", "src/old/Rename.java", "src/Rename.java"),
+                Arguments.of("1 2 src/{old => new}/Rename.java", "src/old/Rename.java", "src/new/Rename.java"),
+                Arguments.of("1\t2\tRename.java => new/Rename.java", "Rename.java", "new/Rename.java"),
+                Arguments.of("1\t2\tRenameOld.java => RenameNew.java", "RenameOld.java", "RenameNew.java")
             )
-        }
     }
 
-    @Test
-    fun isFileline() {
-        assertThat(GitLogNumstatParsingHelper.isFileLine(fileLine)).isTrue()
+    @ParameterizedTest
+    @MethodSource("data")
+    fun isFileline(fileLine: String) {
+        assertTrue(GitLogNumstatParsingHelper.isFileLine(fileLine))
     }
 
-    @Test
-    fun parseModification() {
+    @ParameterizedTest
+    @MethodSource("data")
+    fun parseModification(fileLine: String, oldFilename: String, newFilename: String) {
         val modification = GitLogNumstatParsingHelper.parseModification(fileLine)
 
         assertThat(modification).extracting(