+++ analysis/import/GitLogParser/src/test/kotlin/de/maibornwolff/codecharta/importer/gitlogparser/parser/git/helper/GitLogRawParsingHelperTest.kt
@@ -2,7 +2,7 @@ package de.maibornwolff.codecharta.importer.gitlogparser.parser.git.helper
 
 import de.maibornwolff.codecharta.importer.gitlogparser.input.Modification
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 class GitLogRawParsingHelperTest {